requirements/common.txt
asgi-redis==0.14.1
channels==0.17.2
commonmark==0.5.4
daphne==0.15.0
django==1.10.2
django-import-export==0.5.1
djangorestframework==3.5.2
djangorestframework-expiring-authtoken==0.1.4
django-cors-headers==1.3.1
django-rest-auth[with_social]
drfdocs==0.0.11
pika==0.10.0
pickleshare==0.7.4
Pillow==3.4.2
#psycopg2==2.6.2 #==> for Postgres@9.6
psycopg2==2.7.3.2 #==> for Postgres 10
PyYaml==3.12
proc==0.10.1
rstr==2.2.5
